Make library compatible with callbacks
The release of 1.0.0 broke backwards compatibility
by replacing all callbacks with Promises; this
change restores the ability to use all previous
functions like callbacks and retains the ability
of 1.0.0 to use these functions like Promises.

Minor fixes to tests which have broken following
changes in metadata at the given urls.
