+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "html-metadata",
-  "version": "1.0.0",
+  "version": "1.1.0",
   "description": "Scrapes metadata of several different standards",
   "main": "index.js",
   "dependencies": {
